+++ taglib/src/main/java/org/apache/struts/taglib/bean/DefineTag.java
@@ -159,9 +159,8 @@ public class DefineTag extends BodyTagSupport {
      *
      * @throws JspException if a JSP exception has occurred
      */
-    @SuppressWarnings("deprecation")
     public int doStartTag() throws JspException {
-        return (EVAL_BODY_TAG);
+        return (EVAL_BODY_BUFFERED);
     }
 
     /**