+++ taglib/src/main/java/org/apache/struts/taglib/html/BaseInputTag.java
@@ -210,9 +210,8 @@ public abstract class BaseInputTag extends BaseHandlerTag {
      *
      * @throws JspException if a JSP exception has occurred
      */
-    @SuppressWarnings("deprecation")
     public int doStartTag() throws JspException {
-        return (EVAL_BODY_TAG);
+        return (EVAL_BODY_BUFFERED);
     }
 
     /**