+++ taglib/src/main/java/org/apache/struts/taglib/html/LinkTag.java
@@ -311,13 +311,12 @@ public class LinkTag extends BaseHandlerTag {
      *
      * @throws JspException if a JSP exception has occurred
      */
-    @SuppressWarnings("deprecation")
     public int doStartTag() throws JspException {
         this.text = null;
         this.parameters.clear();
 
         // Evaluate the body of this tag
-        return (EVAL_BODY_TAG);
+        return (EVAL_BODY_BUFFERED);
     }
 
     /**