+++ taglib/src/main/java/org/apache/struts/taglib/html/MultiboxTag.java
@@ -123,12 +123,11 @@ public class MultiboxTag extends BaseHandlerTag {
      *
      * @throws JspException if a JSP exception has occurred
      */
-    @SuppressWarnings("deprecation")
     public int doStartTag() throws JspException {
         // Defer processing until the end of this tag is encountered
         this.constant = null;
 
-        return (EVAL_BODY_TAG);
+        return (EVAL_BODY_BUFFERED);
     }
 
     /**