+++ taglib/src/main/java/org/apache/struts/taglib/html/ParamTag.java
@@ -79,9 +79,8 @@ public class ParamTag extends BodyTagSupport {
     /**
      * @throws JspException if a JSP exception has occurred
      */
-    @SuppressWarnings("deprecation")
     public int doStartTag() throws JspException {
-        return (EVAL_BODY_TAG);
+        return (EVAL_BODY_BUFFERED);
     }
 
     /**