+++ taglib/src/main/java/org/apache/struts/taglib/html/SelectTag.java
@@ -170,7 +170,6 @@ public class SelectTag extends BaseHandlerTag {
      *
      * @throws JspException if a JSP exception has occurred
      */
-    @SuppressWarnings("deprecation")
     public int doStartTag() throws JspException {
         TagUtils.getInstance().write(pageContext, renderSelectStartElement());
 
@@ -179,7 +178,7 @@ public class SelectTag extends BaseHandlerTag {
 
         this.calculateMatchValues();
 
-        return (EVAL_BODY_TAG);
+        return (EVAL_BODY_BUFFERED);
     }
 
     /**