+++ taglib/src/main/java/org/apache/struts/taglib/html/SubmitTag.java
@@ -99,12 +99,11 @@ public class SubmitTag extends BaseHandlerTag {
      *
      * @throws JspException if a JSP exception has occurred
      */
-    @SuppressWarnings("deprecation")
     public int doStartTag() throws JspException {
         // Do nothing until doEndTag() is called
         this.text = null;
 
-        return (EVAL_BODY_TAG);
+        return (EVAL_BODY_BUFFERED);
     }
 
     /**