+++ taglib/src/main/java/org/apache/struts/taglib/html/TextareaTag.java
@@ -46,11 +46,10 @@ public class TextareaTag extends BaseInputTag {
      *
      * @throws JspException if a JSP exception has occurred
      */
-    @SuppressWarnings("deprecation")
     public int doStartTag() throws JspException {
         TagUtils.getInstance().write(pageContext, this.renderTextareaElement());
 
-        return (EVAL_BODY_TAG);
+        return (EVAL_BODY_BUFFERED);
     }
 
     /**